requirements.txt
# VLM-Gym Requirements
# Python 3.11.11 | CUDA 12.1

# Deep Learning
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
transformers==4.57.1
accelerate==1.6.0
safetensors==0.5.3
tokenizers==0.22.1

# VLM APIs
openai==1.77.0
anthropic==0.50.0
google-genai==1.13.0
google-api-python-client==2.169.0
google-auth==2.39.0

# RL / Simulation
mujoco==3.3.2
pettingzoo==1.25.0
Farama-Notifications==0.0.4
pygame==2.6.1
pyglet==1.5.31

# 3D / Rendering
open3d==0.19.0
pyrender==0.1.45
trimesh==4.5.2
objaverse==0.0.6
glfw==2.9.0
PyOpenGL>=3.1.0   
PyOpenGL-accelerate
freetype-py==2.5.1
imageio==2.37.0
pycollada==0.9
manifold3d==3.1.1
mapbox_earcut==1.0.3
embreex==2.17.7.post6
vhacdx==0.0.8.post2
pydracogltf==0.0.4

# Panda3D (for some environments)
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
ursina==7.0.0

# Scientific Computing
numpy==1.26.4
scipy==1.15.2
pandas==2.2.3
scikit-learn==1.7.2
scikit-image==0.25.2
matplotlib==3.10.1
opencv-python==4.11.0.86

# Image Processing
pillow
qwen-vl-utils==0.0.11

# Utilities
tqdm==4.67.1
pyyaml==6.0.2
requests==2.32.3
httpx==0.28.1
huggingface-hub==0.36.0
filelock
joblib==1.5.2
cloudpickle==3.1.1
ConfigArgParse==1.7.1
loguru==0.7.3
colorlog==6.9.0
einops==0.8.1
pyquaternion==0.9.9
shapely==2.1.1
rtree==1.4.0
networkx
lxml==5.4.0
beautifulsoup4==4.13.3
gdown==5.2.0
pycocotools==2.0.8
lvis==0.5.3

# Visualization
plotly==6.3.1
dash==3.2.0
ipywidgets==8.1.7

# Development
Cython==3.0.12
pybind11

# add jupyter notebook for running .ipynb files
notebook==6.5.4
ipykernel==6.28.0
